+++ Jantimizer/DatabaseConnector/DatabaseParserConnector.cs
@@ -6,15 +6,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector.Models
+namespace DatabaseConnector
 {
-    public class DatabaseCommunicator
+    public class DatabaseParserConnector : IDatabaseParserConnector
     {
         public string Name { get; set; }
         public IDbConnector Connector { get; set; }
         public IPlanParser Parser { get; set; }
 
-        public DatabaseCommunicator(string name, IDbConnector connector, IPlanParser parser)
+        public DatabaseParserConnector(string name, IDbConnector connector, IPlanParser parser)
         {
             Name = name;
             Connector = connector;
