+++ Jantimizer/Jantimizer.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Histograms", "Histograms\Hi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatabaseConnector", "DatabaseConnector\DatabaseConnector.csproj", "{73128FA7-3351-4E13-BC2F-EEEFC44DC097}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueryPlanParser", "QueryPlanParser\QueryPlanParser.csproj", "{B77161E7-E29E-4A79-A1DE-850745936198}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{73128FA7-3351-4E13-BC2F-EEEFC44DC097}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{73128FA7-3351-4E13-BC2F-EEEFC44DC097}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{73128FA7-3351-4E13-BC2F-EEEFC44DC097}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B77161E7-E29E-4A79-A1DE-850745936198}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B77161E7-E29E-4A79-A1DE-850745936198}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B77161E7-E29E-4A79-A1DE-850745936198}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B77161E7-E29E-4A79-A1DE-850745936198}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
